+++ includes/class-disable-blog-functions.php
@@ -45,8 +45,64 @@ class Disable_Blog_Functions {
 			return;
 		}
 
+		/**
+		 * Should we pass url query string in the redirect?
+		 *
+		 * Deault is false.
+		 *
+		 * @since 0.5.0
+		 * @param bool $bool true to allow query strings on redirect.
+		 * @return bool
+		 */
+		if ( apply_filters( 'dwpb_pass_query_string_on_redirect', false ) ) {
+			$redirect_url = $this->parse_query_string( $redirect_url );
+		}
+
 		wp_safe_redirect( esc_url_raw( $redirect_url ), $this->get_redirect_status_code( $current_url, $redirect_url ) );
 		exit;
+
+	}
+
+	/**
+	 * Parse the current query string and add it, clean and filter, to the url.
+	 *
+	 * @since 0.5.0
+	 * @param string $url the url any query string will be added to.
+	 * @return string
+	 */
+	private function parse_query_string( $url ) {
+
+		if ( ! isset( $_SERVER['QUERY_STRING'] ) || empty( $_SERVER['QUERY_STRING'] ) ) {
+			return $url;
+		}
+
+		// Setup an array of the current query string variables.
+		$query_vars = array();
+		wp_parse_str( $_SERVER['QUERY_STRING'], $query_vars ); // phpcs:ignore
+
+		/**
+		 * Filter for allowed queary string variables.
+		 *
+		 * @since 0.5.0
+		 * @param array $allowed_query_vars an array of the allowed query variable keys.
+		 * @return array
+		 */
+		$allowed_query_vars = apply_filters( 'dwpb_allowed_query_vars', array() );
+		if ( ! empty( $allowed_query_vars ) && is_array( $allowed_query_vars ) ) {
+			$allowed_query_vars = array_filter( $allowed_query_vars, 'esc_html' );
+			$query_vars         = array_intersect_key( $query_vars, array_flip( $allowed_query_vars ) );
+		}
+
+		// Escaping and sanitization are important.
+		$query_vars = array_filter( $query_vars, 'esc_html' );
+		$query_vars = array_filter( $query_vars, 'esc_html', ARRAY_FILTER_USE_KEY );
+
+		// if we have any query variables, add it to the url.
+		if ( ! empty( $query_vars ) && is_array( $query_vars ) ) {
+			$url = add_query_arg( $query_vars, $url );
+		}
+
+		return $url;
 
 	}